Move factfilter code out of core package
