--- core/src/main/scala/com/rallyhealth/vapors/core/dsl/factfilter/package.scala
@@ -1,3 +1,0 @@
-package com.rallyhealth.vapors.core.dsl
-
-package object factfilter extends Dsl
