+++ core/src/main/scala/com/rallyhealth/vapors/core/evaluator/EvalLoop.scala
@@ -3,7 +3,7 @@ package com.rallyhealth.vapors.core.evaluator
 import cats.instances.function._
 import cats.~>
 import com.rallyhealth.vapors.core.algebra._
-import com.rallyhealth.vapors.core.dsl.Exp
+import com.rallyhealth.vapors.factfilter.dsl.Exp
 
 private[evaluator] final class EvalLoop[T] extends (ExpAlg[T, *] ~> (T => *)) {
