+++ core/src/main/scala/com/rallyhealth/vapors/core/evaluator/package.scala
@@ -1,10 +1,12 @@
 package com.rallyhealth.vapors.core
 
+import cats.free.FreeApplicative
+import com.rallyhealth.vapors.core.algebra.ExpAlg
+
 package object evaluator {
   import cats.instances.function._
-  import dsl._
 
-  def eval[T, A](input: T)(exp: Exp[T, A]): A = {
+  def eval[T, A](input: T)(exp: FreeApplicative[ExpAlg[T, *], A]): A = {
     exp.foldMap(new EvalLoop[T]).apply(input)
   }
