+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/data/FactType.scala
@@ -1,6 +1,7 @@
-package com.rallyhealth.vapors.core.data
+package com.rallyhealth.vapors.factfilter.data
 
 import cats.Order
+import com.rallyhealth.vapors.core.data.NamedLens
 
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe.TypeTag
