+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/data/FactTypeSet.scala
@@ -1,4 +1,4 @@
-package com.rallyhealth.vapors.core.data
+package com.rallyhealth.vapors.factfilter.data
 
 import cats.data.{NonEmptyList, NonEmptyMap, NonEmptySet}
