+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/Facts.scala
@@ -1,0 +1,12 @@
+package com.rallyhealth.vapors.factfilter.dsl
+
+import cats.data.NonEmptyList
+import com.rallyhealth.vapors.factfilter.data.{Fact, Facts}
+
+object Facts {
+
+  def apply(
+    head: Fact,
+    tail: Fact*,
+  ): Facts = NonEmptyList.of(head, tail: _*)
+}
