+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/LogicalOps.scala
@@ -1,6 +1,5 @@
-package com.rallyhealth.vapors.core.dsl.factfilter
+package com.rallyhealth.vapors.factfilter.dsl
 
-import com.rallyhealth.vapors.core.dsl
 import com.rallyhealth.vapors.core.logic.{Intersect, Union}
 
 final class LogicalOps[T, A](private val exp: Exp[T, A]) extends AnyVal {
@@ -8,20 +7,20 @@ final class LogicalOps[T, A](private val exp: Exp[T, A]) extends AnyVal {
   /**
     * Builds an AND expression using infix notation.
     */
-  def and(o: Exp[T, A])(implicit A: Intersect[A]): Exp[T, A] = dsl.and(exp, o)
+  def and(o: Exp[T, A])(implicit A: Intersect[A]): Exp[T, A] = __.and(exp, o)
 
   /**
     * Same as [[and]] but with higher precedent.
     */
-  def &&(o: Exp[T, A])(implicit A: Intersect[A]): Exp[T, A] = dsl.and(exp, o)
+  def &&(o: Exp[T, A])(implicit A: Intersect[A]): Exp[T, A] = __.and(exp, o)
 
   /**
     * Builds an OR expression using infix notation.
     */
-  def or(o: Exp[T, A])(implicit A: Union[A]): Exp[T, A] = dsl.or(exp, o)
+  def or(o: Exp[T, A])(implicit A: Union[A]): Exp[T, A] = __.or(exp, o)
 
   /**
     * Same as [[or]] but with higher precedent.
     */
-  def ||(o: Exp[T, A])(implicit A: Union[A]): Exp[T, A] = dsl.or(exp, o)
+  def ||(o: Exp[T, A])(implicit A: Union[A]): Exp[T, A] = __.or(exp, o)
 }
