+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/Syntax.scala
@@ -1,21 +1,6 @@
-package com.rallyhealth.vapors.core.dsl.factfilter
-
-import com.rallyhealth.vapors.core.dsl
+package com.rallyhealth.vapors.factfilter.dsl
 
 private[dsl] trait Syntax {
-  self: Dsl =>
-
-  /**
-    * An alias to this [[dsl]] object, so you can use infix operators and more easily explore
-    * the list of supported expression builders.
-    *
-    * Example:
-    * {{{
-    *   __.withFactsOfType(FactTypes.Age)
-    *     .whereAnyValue(__ > 35)
-    * }}}
-    */
-  final val __ = this
 
   def >[T : Ordering](lowerBound: T): CondExp[T] = greaterThan(lowerBound)
   def >=[T : Ordering](lowerBound: T): CondExp[T] = greaterThanOrEqual(lowerBound)
